+++ CHANGELOG.md
@@ -38,9 +38,11 @@
   `--fakeroot`, for example).
 - The `remote status` command will now print the username, realname, and email
   of the logged-in user, if available.
-- OCI-mode now supports `--overlay <dir>` flag, allowing writes to the
-  filesystem to persist across runs of the OCI container. Multiple overlays can be
-  specified, but all but one must be read-only (`--overlay <dir>:ro`).
+- OCI-mode now supports an `--overlay <arg>` flag. `<arg>` can be a writable
+  directory, in which case changes to the filesystem will persist across runs of
+  the OCI container. Alternatively, `<arg>` can be `<dir>:ro` or the path of a
+  squashfs image, either of which will be mounted as a read-only overlay.
+  Multiple overlays can be specified, but all but one must be read-only.
 - The `tap` CNI plugin, new to github.com/containernetworking/plugins v1.3.0,
   is now provided.
 - OCI-mode now supports the `--workdir <workdir>` option. If this option is